requirements.txt
requests==2.32.3
bs4==0.0.2
lxml==5.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.8
flower==2.0.1



